storage: extract Key/Value/KvPair to keys crate
