+++ components/keys/Cargo.toml
@@ -10,5 +10,7 @@ derive_more = "0.15.0"
 hex = "0.3"
 kvproto = { git = "https://github.com/pingcap/kvproto.git" }
 
+tikv_util = { path = "../tikv_util" }
+
 [dev-dependencies]
 panic_hook = { path = "../panic_hook" }
